+++ src/global/service.go
@@ -1,0 +1,18 @@
+// Copyright 2025 BackendServerTemplate Authors. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+package global
+
+import resource "github.com/SongZihuan/BackendServerTemplate"
+
+var ServiceConfig = resource.ServiceConfig
+
+const (
+	Args1Install    = resource.Args1Install
+	Args1Uninstall1 = resource.Args1Uninstall1
+	Args1Uninstall2 = resource.Args1Uninstall2
+	Args1Start      = resource.Args1Start
+	Args1Stop       = resource.Args1Stop
+	Args1Restart    = resource.Args1Restart
+)